vms: Make it more imperative

+++ modules/vms/default.nix
@@ -12,31 +12,7 @@ let
     hostCpuOnly = true;
   };
 
-  vmOpts = { name, ... }: {
-    options = {
-      hostName = lib.mkOption {
-        default = name;
-        type = lib.types.str;
-      };
-      memory = lib.mkOption {
-        type = lib.types.str;
-      };
-      diskSize = lib.mkOption {
-        type = lib.types.str;
-      };
-      cpus = lib.mkOption {
-        type = lib.types.int;
-      };
-      os = lib.mkOption {
-        type = lib.types.str;
-      };
-      macAddress = lib.mkOption {
-        default = "";
-        type = lib.types.str;
-      };
-    };
-
-  };
+  firmware = (tinyQemu + "/share/qemu/edk2-x86_64-code.fd");
 
 in
 
@@ -47,9 +23,13 @@ in
       default = false;
       type = lib.types.bool;
     };
+    stateDir = lib.mkOption {
+      default = "/var/lib/guests";
+      type = lib.types.str;
+    };
     guests = lib.mkOption {
-      default = {};
-      type = with lib.types; attrsOf (submodule vmOpts);
+      default = [];
+      type = lib.types.listOf lib.types.str;
     };
   };
 
@@ -59,47 +39,62 @@ in
       { assertion = config.ab.net.bridge.enable; }
     ];
 
+    system.activationScripts.microvm-host = ''
+      mkdir -p ${cfg.stateDir}
+    '';
+
     environment.systemPackages = [
       tinyQemu
-    ];
+      (pkgs.writeScriptBin "install-guest" ''
+        [ "$1" != "" ] || exit 1
+        [ "$2" != "" ] || exit 1
 
-    systemd.services = lib.mapAttrs' (name: guest: lib.nameValuePair "guest-${name}" {
-      wantedBy = [ "multi-user.target" ];
-      script = ''
-        disks=/var/lib/guests/disks
-        mkdir -p "$disks"
-        hda="$disks/${name}.img"
-        vars="$disks/${name}.VARS-x86_64.fd"
-        size="${guest.diskSize}"
-
-        firmware="${tinyQemu}/share/qemu/edk2-x86_64-code.fd"
-
-        sock="/run/guest-${name}.api.sock"
-        rm -f $sock
+        store="${cfg.stateDir}/$1"
+        mkdir -p "$store"
+        hda="$store/hda.img"
+        vars="$store/VARS-x86_64.fd"
 
         # Initialize the associated files
         if ! test -f $hda; then
+          echo Creating associated guest files
           touch $hda
           ${pkgs.e2fsprogs}/bin/chattr +C $hda # Improves performance on btrfs
-          ${pkgs.util-linux}/bin/fallocate -l "$size" "$hda"
+          ${pkgs.util-linux}/bin/fallocate -l "$2" "$hda"
           cat ${tinyQemu}/share/qemu/edk2-i386-vars.fd > $vars # EFI variables store
+          echo "THREADS=1" >> "$store/config"
+          echo "MEMORY=1G" >> "$store/config"
         fi
 
-        # Run the installation process
-        if ! test -f $hda.installed; then
-          ${tinyQemu}/bin/qemu-kvm -drive file=${nixosInstaller},if=virtio,format=raw,media=disk,readonly=on -drive file=$hda,if=virtio,format=raw,media=disk -nic tap,id=net0,ifname="vm-$name",model=virtio,script=no,downscript=no -nographic -vga none -serial file:$hda.console.log -cpu host -m 2G -drive if=pflash,format=raw,unit=0,file=$firmware,readonly=on -drive if=pflash,format=raw,unit=1,file=$vars -monitor unix:$sock,server,nowait
-          touch $hda.installed
+        echo Running the guest with the installation ISO
+        ${tinyQemu}/bin/qemu-kvm -drive file=${nixosInstaller},if=virtio,format=raw,media=disk,readonly=on -drive file=$hda,if=virtio,format=raw,media=disk -nic tap,id=net0,ifname="vm-$name",model=virtio,script=no,downscript=no -nographic -vga none -serial stdio -cpu host -m 2G -drive if=pflash,format=raw,unit=0,file=${firmware},readonly=on -drive if=pflash,format=raw,unit=1,file=$vars -monitor unix:$sock,server,nowait
+      '')
+    ];
+
+    systemd.services = lib.mkMerge ([{"guest@" = {
+      restartIfChanged = false;
+      scriptArgs = "%i";
+      script = (''
+        store="${cfg.stateDir}/$1"
+        if ! [ -e "$store" ]; then
+          echo "Error: The specified host has not been initialized yet."
+          echo "Use 'vm-install name disk-size' to install its OS."
+          exit 1
         fi
 
-        # Run the guest without the installation media present
-        ${tinyQemu}/bin/qemu-kvm -drive file=$hda,if=virtio,format=raw,media=disk -nic tap,id=net0,ifname=vm-${name},model=virtio,script=no,downscript=no -nographic -vga none -serial none -cpu host -smp ${toString guest.cpus} -m ${guest.memory} -drive if=pflash,format=raw,unit=0,file=$firmware,readonly=on -drive if=pflash,format=raw,unit=1,file=$vars -monitor unix:$sock,server,nowait
-      '';
-      preStop = ''
-        echo 'system_powerdown' | ${pkgs.socat}/bin/socat - UNIX-CONNECT:/run/guest-${name}.api.sock
-        sleep 10
-      '';
-    }) cfg.guests;
+        hda="$store/hda.img"
+        vars="$store/VARS-x86_64.fd"
+        sock="/run/guest-$1.api.sock"
+        rm -f $sock
 
-    ab.net.bridge.vmTaps = lib.mapAttrs (name: guest: { inherit (guest) macAddress; }) cfg.guests;
+        . "$store/config"
+
+        ${tinyQemu}/bin/qemu-kvm -drive file=$hda,if=virtio,format=raw,media=disk -nic tap,id=net0,ifname="vm-$1",model=virtio,script=no,downscript=no -nographic -vga none -serial none -cpu host -smp $THREADS -m $MEMORY -drive if=pflash,format=raw,unit=0,file=${firmware},readonly=on -drive if=pflash,format=raw,unit=1,file=$vars -monitor unix:$sock,server,nowait
+      '');
+    };} (builtins.listToAttrs (map (name: lib.nameValuePair "guest@${name}" {
+      wantedBy = [ "multi-user.target" ];
+      restartIfChanged = false;
+      overrideStrategy = "asDropin";
+    }) cfg.guests))]);
+
   };
 }
